+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 
 - added: `git machete git{hub,lab} update-{pr,mr}-descriptions` subcommands
 - added: `git machete gith{hub,lab} create-{pr,mr} --update-related-descriptions` flags
+- added: `machete.traverse.fetch.<remote>` git config key to selectively exclude remotes from `git machete traverse --fetch` (contributed by @gjulianm)
 
 ## New in git-machete 3.30.0